Player Class Implemented
Publically editable Strength,Mass and MaxSpeed variables added. The last two are implemented but Strength needs fully working players that can hit eachother.

Movement function moved to Player.cs
Raycasting is now a parent of Player and returns booleans
Controller is a parent of raycasting
Inheritance structure can be changed if we so wish or if a new script is added to the chain